Adding sweetalert2 and error alerts

+++ resources/views/users/index.blade.php
@@ -34,15 +34,46 @@
 
 $(function(){
  $(".delete").on("click",function(){ 
-      $.ajax({
-      method: "delete",
-      url: "http://localhost/CV/public/users/"+$(this).data("id"),
-      //data: { name: "John", location: "Boston" }
-      })
-      .done(function(response) {
-      location.reload();
-      alert("SUKCES");
-       });
+  Swal.fire({
+    title: "Are you sure?",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Tak",
+    cancelButtonText:"Nie"
+            }).then((result) => {
+  if (result.isConfirmed) {
+     $.ajax({
+     method: "delete",
+     url: "http://localhost/CV/public/users/"+$(this).data("id"),
+     data: { name: "John", location: "Boston" }
+            })
+    .then(
+      function(response) {
+        location.reload();
+        Swal.fire({
+        title: "Deleted!",
+        text: "Your file has been deleted.",
+        icon: "success"
+                });
+                        },
+    function(jqXHR, textStatus, errorThrown) {
+      var responseJSON = JSON.parse(jqXHR.responseText);
+      console.log(responseJSON.message);
+      Swal.fire({
+     title: "error",
+      text: responseJSON.message,
+      icon: "error",
+               });
+                                             }
+          );
+
+   
+                         };
+ 
+                                });
+   
        });
 });
